requirements.txt
# TicketDrop 2.0 DOE Workflow Dependencies
# Install: pip install -r requirements.txt

# Google Sheets integration
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0

# Data processing
pandas>=2.1.0

# Environment management
python-dotenv>=1.0.0

# HTTP requests (for API calls)
requests>=2.31.0

# Streamlit driver app
streamlit>=1.29.0
Pillow>=10.1.0
